+++ CHANGELOG.md
@@ -5,14 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.1] - 2021-06-25
+## [0.9.0] - 2021-06-28
+
+### 🔥 Added
+
+- `distributors list` page
+- `add distributor` page
+- `edit distributor` page
+- `delete distributor` page
+- protect authorized routes
+
+### 💪 Updated
+
+- `create user` page
+
+## [0.8.1] - 2021-06-27
 
 ### 👾 Fixed
 
 - links to changelog
 - redirections to `RFQ list` afrer adding new RFQ
 
-## [0.8.0] - 2021-06-25
+## [0.8.0] - 2021-06-26
 
 ### 🔥 Added